Extracted handlers from tabpy.py to handlers package; updated info method to return version.
