+++ tabpy-server/tabpy_server/handlers/__init__.py
@@ -1,0 +1,11 @@
+from tabpy_server.handlers.base_handler import BaseHandler
+from tabpy_server.handlers.main_handler import MainHandler
+from tabpy_server.handlers.management_handler import ManagementHandler
+
+from tabpy_server.handlers.endpoint_handler import EndpointHandler
+from tabpy_server.handlers.endpoints_handler import EndpointsHandler
+from tabpy_server.handlers.evaluation_plane_handler import EvaluationPlaneHandler
+from tabpy_server.handlers.query_plane_handler import QueryPlaneHandler
+from tabpy_server.handlers.service_info_handler import ServiceInfoHandler
+from tabpy_server.handlers.status_handler import StatusHandler
+from tabpy_server.handlers.upload_destination_handler import UploadDestinationHandler
